Refactor: FeedbackTable more
